feat(settings): Add AI model settings screen and update localization strings
