+++ app/src/main/java/com/nguyenmoclam/tutorialyoutubemadesimple/ui/components/SettingsComponents.kt
@@ -865,779 +865,3 @@ fun SettingsItem(
         }
     }
 }
-
-/**
- * AI Model Settings component that displays AI model configuration options.
- */
-@SuppressLint("DefaultLocale")
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AIModelSettings(
-    openRouterApiKey: String,
-    onApiKeyChanged: (String) -> Unit,
-    selectedModel: String,
-    onModelSelected: (String) -> Unit,
-    currentCredits: Double,
-    isLoading: Boolean = false,
-    validationState: ApiKeyValidationState = ApiKeyValidationState.NOT_VALIDATED,
-    models: List<ModelInfo> = emptyList(),
-    onRefreshModels: () -> Unit = {},
-    onApplyFilter: (ModelFilter.Category, String) -> Unit = { _, _ -> },
-    onClearFilter: (ModelFilter.Category) -> Unit = { _ -> },
-    onSetSortOption: (ModelFilter.SortOption) -> Unit = { _ -> },
-    currentFilters: Map<ModelFilter.Category, Set<String>> = emptyMap(),
-    currentSortOption: ModelFilter.SortOption = ModelFilter.SortOption.TOP_WEEKLY,
-    onLoadMoreModels: () -> Unit = {},
-    hasMoreModels: Boolean = false,
-    // New parameters for credits/usage monitoring
-    creditStatusState: UsageViewModel.CreditStatusState = UsageViewModel.CreditStatusState.Loading,
-    tokenUsageSummaryState: UsageViewModel.TokenUsageSummaryState = UsageViewModel.TokenUsageSummaryState.Loading,
-    selectedTimeRange: UsageViewModel.TimeRange = UsageViewModel.TimeRange.LAST_30_DAYS,
-    onTimeRangeSelected: (UsageViewModel.TimeRange) -> Unit = {},
-    onRefreshCredits: () -> Unit = {},
-    onViewAllModelsClick: () -> Unit = {}
-) {
-    // State for password visibility
-    var passwordVisible by remember { mutableStateOf(false) }
-    
-    // Clipboard manager for paste functionality
-    val clipboardManager = LocalClipboardManager.current
-
-    Column(modifier = Modifier.fillMaxWidth()) {
-        // API Key Section
-        Text(
-            stringResource(R.string.api_key_label),
-            fontWeight = FontWeight.Medium
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(
-            stringResource(R.string.api_key_description),
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        // API key input field with password visibility toggle
-        OutlinedTextField(
-            value = openRouterApiKey,
-            onValueChange = {
-                onApiKeyChanged(it)
-            },
-            modifier = Modifier.fillMaxWidth(),
-            label = { Text(stringResource(R.string.api_key_label)) },
-            placeholder = { Text(stringResource(R.string.api_key_hint)) },
-            visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-            singleLine = true,
-            trailingIcon = {
-                Row {
-                    // Paste button
-                    IconButton(onClick = {
-                        clipboardManager.getText()?.text?.let {
-                            onApiKeyChanged(it)
-                        }
-                    }) {
-                        Icon(
-                            imageVector = Icons.Default.ContentPaste,
-                            contentDescription = stringResource(R.string.api_key_paste)
-                        )
-                    }
-
-                    // Password visibility toggle
-                    IconButton(onClick = { passwordVisible = !passwordVisible }) {
-                        Icon(
-                            imageVector = if (passwordVisible) Icons.Default.VisibilityOff else Icons.Default.Visibility,
-                            contentDescription = stringResource(R.string.api_key_toggle_visibility)
-                        )
-                    }
-                }
-            },
-            // Show validation indicator
-            leadingIcon = when (validationState) {
-                ApiKeyValidationState.VALID -> {
-                    {
-                        Icon(
-                            imageVector = Icons.Default.CheckCircle,
-                            contentDescription = "Valid",
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                }
-                ApiKeyValidationState.INVALID, ApiKeyValidationState.INVALID_FORMAT -> {
-                    {
-                        Icon(
-                            imageVector = Icons.Default.Error,
-                            contentDescription = "Invalid",
-                            tint = MaterialTheme.colorScheme.error
-                        )
-                    }
-                }
-                ApiKeyValidationState.VALIDATING -> {
-                    {
-                        CircularProgressIndicator(
-                            modifier = Modifier
-                                .size(24.dp)
-                                .padding(4.dp),
-                            strokeWidth = 2.dp
-                        )
-                    }
-                }
-                else -> null
-            },
-            // Set colors based on validation state
-            colors = TextFieldDefaults.outlinedTextFieldColors(
-                focusedBorderColor = when (validationState) {
-                    ApiKeyValidationState.VALID -> MaterialTheme.colorScheme.primary
-                    ApiKeyValidationState.INVALID, ApiKeyValidationState.INVALID_FORMAT -> MaterialTheme.colorScheme.error
-                    else -> MaterialTheme.colorScheme.primary
-                },
-                unfocusedBorderColor = when (validationState) {
-                    ApiKeyValidationState.VALID -> MaterialTheme.colorScheme.primary.copy(alpha = 0.5f)
-                    ApiKeyValidationState.INVALID, ApiKeyValidationState.INVALID_FORMAT -> MaterialTheme.colorScheme.error.copy(alpha = 0.5f)
-                    else -> MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.3f)
-                }
-            ),
-            supportingText = {
-                Text(
-                    text = when (validationState) {
-                        ApiKeyValidationState.INVALID_FORMAT -> stringResource(R.string.api_key_format_error)
-                        ApiKeyValidationState.INVALID -> stringResource(R.string.api_key_validation_error)
-                        ApiKeyValidationState.VALID -> stringResource(R.string.api_key_validation_success)
-                        ApiKeyValidationState.VALIDATING -> stringResource(R.string.api_key_validating)
-                        else -> stringResource(R.string.api_key_helper)
-                    },
-                    color = when (validationState) {
-                        ApiKeyValidationState.INVALID_FORMAT, ApiKeyValidationState.INVALID -> MaterialTheme.colorScheme.error
-                        ApiKeyValidationState.VALID -> MaterialTheme.colorScheme.primary
-                        else -> MaterialTheme.colorScheme.onSurfaceVariant
-                    }
-                )
-            }
-        )
-
-        // Status Indicator (Using Card instead of Chip)
-        if (validationState != ApiKeyValidationState.NOT_VALIDATED) {
-            Spacer(modifier = Modifier.height(8.dp))
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.End
-            ) {
-                Card(
-                    modifier = Modifier.padding(4.dp),
-                    colors = CardDefaults.cardColors(
-                        containerColor = when (validationState) {
-                            ApiKeyValidationState.VALID -> MaterialTheme.colorScheme.primaryContainer
-                            ApiKeyValidationState.INVALID, ApiKeyValidationState.INVALID_FORMAT -> MaterialTheme.colorScheme.errorContainer
-                            ApiKeyValidationState.VALIDATING -> MaterialTheme.colorScheme.secondaryContainer
-                            else -> MaterialTheme.colorScheme.surfaceVariant
-                        }
-                    )
-                ) {
-                    Row(
-                        modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        // Leading icon
-                        when (validationState) {
-                            ApiKeyValidationState.VALID -> {
-                                Icon(
-                                    imageVector = Icons.Default.CheckCircle,
-                                    contentDescription = null,
-                                    tint = MaterialTheme.colorScheme.primary,
-                                    modifier = Modifier.size(16.dp)
-                                )
-                            }
-                            ApiKeyValidationState.INVALID, ApiKeyValidationState.INVALID_FORMAT -> {
-                                Icon(
-                                    imageVector = Icons.Default.Error,
-                                    contentDescription = null,
-                                    tint = MaterialTheme.colorScheme.error,
-                                    modifier = Modifier.size(16.dp)
-                                )
-                            }
-                            ApiKeyValidationState.VALIDATING -> {
-                                CircularProgressIndicator(
-                                    modifier = Modifier
-                                        .size(16.dp)
-                                        .padding(2.dp),
-                                    strokeWidth = 2.dp
-                                )
-                            }
-                            else -> {}
-                        }
-
-                        Spacer(modifier = Modifier.width(8.dp))
-
-                        // Text
-                        Text(
-                            text = when (validationState) {
-                                ApiKeyValidationState.VALID -> stringResource(R.string.api_key_validation_success)
-                                ApiKeyValidationState.INVALID -> stringResource(R.string.api_key_validation_error)
-                                ApiKeyValidationState.INVALID_FORMAT -> stringResource(R.string.api_key_format_error)
-                                ApiKeyValidationState.VALIDATING -> stringResource(R.string.api_key_validating)
-                                else -> ""
-                            },
-                            style = MaterialTheme.typography.labelMedium
-                        )
-                    }
-                }
-            }
-        }
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-        // Credits and Usage Monitoring Section - replacing the old credits display
-        if (validationState == ApiKeyValidationState.VALID) {
-            CreditAndUsageMonitoring(
-                creditStatusState = creditStatusState,
-                tokenUsageSummaryState = tokenUsageSummaryState,
-                selectedTimeRange = selectedTimeRange,
-                onTimeRangeSelected = onTimeRangeSelected,
-                onRefreshCredits = onRefreshCredits,
-                onViewAllModelsClick = onViewAllModelsClick
-            )
-            
-            Spacer(modifier = Modifier.height(24.dp))
-        }
-
-        // Model Selection Section
-        Text(
-            stringResource(R.string.select_ai_model),
-            fontWeight = FontWeight.Medium
-        )
-        
-        Spacer(modifier = Modifier.height(16.dp))
-        
-        if (validationState == ApiKeyValidationState.VALID) {
-            // Use our new ModelSelectionComponent
-            ModelSelectionComponent(
-                models = models,
-                selectedModelId = selectedModel,
-                onModelSelected = onModelSelected,
-                currentFilters = currentFilters,
-                currentSortOption = currentSortOption,
-                onApplyFilter = onApplyFilter,
-                onClearFilter = onClearFilter,
-                onSetSortOption = onSetSortOption,
-                isLoading = isLoading,
-                onRefresh = onRefreshModels,
-                onLoadMore = onLoadMoreModels,
-                hasMoreModels = hasMoreModels
-            )
-        } else {
-            // Show message requiring valid API key
-            ApiKeyRequiredMessage()
-        }
-
-        // Help section with OpenRouter guidance
-        Spacer(modifier = Modifier.height(24.dp))
-        OpenRouterHelpSection()
-    }
-}
-
-/**
- * Component for displaying credit balance and token usage monitoring.
- * Shows current credits, usage status, and detailed statistics.
- */
-@SuppressLint("DefaultLocale")
-@Composable
-fun CreditAndUsageMonitoring(
-    creditStatusState: UsageViewModel.CreditStatusState,
-    tokenUsageSummaryState: UsageViewModel.TokenUsageSummaryState,
-    selectedTimeRange: UsageViewModel.TimeRange,
-    onTimeRangeSelected: (UsageViewModel.TimeRange) -> Unit,
-    onRefreshCredits: () -> Unit,
-    onViewAllModelsClick: () -> Unit = {}
-) {
-    Column(modifier = Modifier.fillMaxWidth()) {
-        // Header
-        Text(
-            text = stringResource(R.string.credits_and_usage),
-            fontWeight = FontWeight.Medium
-        )
-        
-        Spacer(modifier = Modifier.height(16.dp))
-        
-        // Credit Status Card
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 8.dp),
-            colors = CardDefaults.cardColors(
-                containerColor = when (creditStatusState) {
-                    is UsageViewModel.CreditStatusState.Success -> {
-                        when (creditStatusState.creditStatus.status) {
-                            CreditStatus.BalanceStatus.CRITICAL -> MaterialTheme.colorScheme.errorContainer
-                            CreditStatus.BalanceStatus.LOW -> MaterialTheme.colorScheme.secondaryContainer
-                            CreditStatus.BalanceStatus.OK -> MaterialTheme.colorScheme.tertiaryContainer
-                        }
-                    }
-                    else -> MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-                }
-            ),
-            elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp)
-            ) {
-                Column {
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(
-                            text = stringResource(R.string.available_credits),
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.SemiBold
-                        )
-                        
-                        // Refresh button
-                        IconButton(
-                            onClick = onRefreshCredits,
-                            modifier = Modifier.size(36.dp)
-                        ) {
-                            when (creditStatusState) {
-                                is UsageViewModel.CreditStatusState.Loading -> {
-                                    CircularProgressIndicator(
-                                        modifier = Modifier.size(20.dp),
-                                        strokeWidth = 2.dp
-                                    )
-                                }
-                                else -> {
-                                    Icon(
-                                        imageVector = Icons.Default.Refresh,
-                                        contentDescription = stringResource(R.string.refresh_credits),
-                                        tint = MaterialTheme.colorScheme.onSurfaceVariant
-                                    )
-                                }
-                            }
-                        }
-                    }
-                    
-                    Spacer(modifier = Modifier.height(16.dp))
-                    
-                    when (creditStatusState) {
-                        is UsageViewModel.CreditStatusState.Loading -> {
-                            Row(
-                                verticalAlignment = Alignment.CenterVertically,
-                                horizontalArrangement = Arrangement.Center,
-                                modifier = Modifier.fillMaxWidth()
-                            ) {
-                                CircularProgressIndicator(
-                                    modifier = Modifier.size(24.dp),
-                                    strokeWidth = 2.dp
-                                )
-                                Spacer(modifier = Modifier.width(8.dp))
-                                Text(
-                                    text = stringResource(R.string.loading_credits),
-                                    style = MaterialTheme.typography.bodyMedium
-                                )
-                            }
-                        }
-                        is UsageViewModel.CreditStatusState.Success -> {
-                            val creditStatus = creditStatusState.creditStatus
-                            
-                            // Credit amount
-                            Text(
-                                text = "$${String.format("%.2f", creditStatus.credits)}",
-                                style = MaterialTheme.typography.headlineMedium,
-                                fontWeight = FontWeight.Bold,
-                                color = when (creditStatus.status) {
-                                    CreditStatus.BalanceStatus.CRITICAL -> MaterialTheme.colorScheme.error
-                                    CreditStatus.BalanceStatus.LOW -> MaterialTheme.colorScheme.secondary
-                                    CreditStatus.BalanceStatus.OK -> MaterialTheme.colorScheme.tertiary
-                                }
-                            )
-                            
-                            Spacer(modifier = Modifier.height(4.dp))
-                            
-                            // Status message
-                            val statusMessage = when (creditStatus.status) {
-                                CreditStatus.BalanceStatus.CRITICAL -> stringResource(
-                                    R.string.critical_balance_warning,
-                                    String.format("%.2f", creditStatus.credits)
-                                )
-                                CreditStatus.BalanceStatus.LOW -> stringResource(
-                                    R.string.low_balance_warning,
-                                    String.format("%.2f", creditStatus.credits)
-                                )
-                                CreditStatus.BalanceStatus.OK -> stringResource(
-                                    R.string.credit_balance_status,
-                                    String.format("%.2f", creditStatus.credits)
-                                )
-                            }
-                            
-                            Text(
-                                text = statusMessage,
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                            
-                            Spacer(modifier = Modifier.height(16.dp))
-                            
-                            // Credit details
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                horizontalArrangement = Arrangement.SpaceBetween
-                            ) {
-                                Column {
-                                    Text(
-                                        text = stringResource(R.string.credit_granted),
-                                        style = MaterialTheme.typography.bodySmall,
-                                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                                    )
-                                    Text(
-                                        text = "$${String.format("%.2f", creditStatus.creditGranted)}",
-                                        style = MaterialTheme.typography.bodyMedium,
-                                        fontWeight = FontWeight.Medium
-                                    )
-                                }
-                                
-                                Column(horizontalAlignment = Alignment.End) {
-                                    Text(
-                                        text = stringResource(R.string.credit_used),
-                                        style = MaterialTheme.typography.bodySmall,
-                                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                                    )
-                                    Text(
-                                        text = "$${String.format("%.2f", creditStatus.creditUsed)}",
-                                        style = MaterialTheme.typography.bodyMedium,
-                                        fontWeight = FontWeight.Medium
-                                    )
-                                }
-                            }
-                            
-                            Spacer(modifier = Modifier.height(8.dp))
-                            
-                            // Last updated
-                            Text(
-                                text = stringResource(
-                                    R.string.last_updated,
-                                    formatLastUpdated(creditStatus.lastUpdated)
-                                ),
-                                style = MaterialTheme.typography.labelSmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.7f),
-                                modifier = Modifier.fillMaxWidth(),
-                                textAlign = TextAlign.End
-                            )
-                        }
-                        is UsageViewModel.CreditStatusState.Error -> {
-                            Column(
-                                horizontalAlignment = Alignment.CenterHorizontally,
-                                modifier = Modifier.fillMaxWidth()
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.Error,
-                                    contentDescription = null,
-                                    tint = MaterialTheme.colorScheme.error,
-                                    modifier = Modifier.size(40.dp)
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-                                Text(
-                                    text = stringResource(R.string.credit_fetch_error),
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    color = MaterialTheme.colorScheme.error
-                                )
-                                Spacer(modifier = Modifier.height(4.dp))
-                                Text(
-                                    text = creditStatusState.message,
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(24.dp))
-        
-        // Token Usage Section
-        Text(
-            text = stringResource(R.string.token_usage),
-            fontWeight = FontWeight.Medium
-        )
-        
-        Spacer(modifier = Modifier.height(8.dp))
-        
-        // Time range selector
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            // Time range chips
-            listOf(
-                UsageViewModel.TimeRange.LAST_7_DAYS to stringResource(R.string.last_7_days),
-                UsageViewModel.TimeRange.LAST_30_DAYS to stringResource(R.string.last_30_days),
-                UsageViewModel.TimeRange.ALL_TIME to stringResource(R.string.all_time)
-            ).forEach { (range, label) ->
-                val isSelected = selectedTimeRange == range
-                
-                FilterChip(
-                    selected = isSelected,
-                    onClick = { onTimeRangeSelected(range) },
-                    label = { Text(label) },
-                    leadingIcon = if (isSelected) {
-                        {
-                            Icon(
-                                imageVector = Icons.Default.Check,
-                                contentDescription = null,
-                                modifier = Modifier.size(16.dp)
-                            )
-                        }
-                    } else null
-                )
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(16.dp))
-        
-        // Token usage summary
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 8.dp),
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-            ),
-            elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp)
-            ) {
-                when (tokenUsageSummaryState) {
-                    is UsageViewModel.TokenUsageSummaryState.Loading -> {
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            horizontalArrangement = Arrangement.Center,
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(vertical = 16.dp)
-                        ) {
-                            CircularProgressIndicator(
-                                modifier = Modifier.size(24.dp),
-                                strokeWidth = 2.dp
-                            )
-                            Spacer(modifier = Modifier.width(8.dp))
-                            Text(
-                                text = stringResource(R.string.loading_usage_data),
-                                style = MaterialTheme.typography.bodyMedium
-                            )
-                        }
-                    }
-                    is UsageViewModel.TokenUsageSummaryState.Success -> {
-                        val summary = tokenUsageSummaryState.summary
-                        
-                        // Total tokens used
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text(
-                                text = stringResource(R.string.total_tokens),
-                                style = MaterialTheme.typography.titleMedium,
-                                fontWeight = FontWeight.SemiBold
-                            )
-                            Text(
-                                text = summary.totalTokens.toString(),
-                                style = MaterialTheme.typography.titleMedium,
-                                fontWeight = FontWeight.Bold,
-                                color = MaterialTheme.colorScheme.primary
-                            )
-                        }
-                        
-                        Spacer(modifier = Modifier.height(16.dp))
-                        
-                        // Token breakdown
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween
-                        ) {
-                            Column {
-                                Text(
-                                    text = stringResource(R.string.prompt_tokens),
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                                )
-                                Text(
-                                    text = summary.totalPromptTokens.toString(),
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    fontWeight = FontWeight.Medium
-                                )
-                            }
-                            
-                            Column(horizontalAlignment = Alignment.End) {
-                                Text(
-                                    text = stringResource(R.string.completion_tokens),
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                                )
-                                Text(
-                                    text = summary.totalCompletionTokens.toString(),
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    fontWeight = FontWeight.Medium
-                                )
-                            }
-                        }
-                        
-                        Spacer(modifier = Modifier.height(16.dp))
-                        
-                        // Total estimated cost
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text(
-                                text = stringResource(R.string.estimated_cost),
-                                style = MaterialTheme.typography.titleSmall,
-                                fontWeight = FontWeight.Medium
-                            )
-                            Text(
-                                text = "$${String.format("%.4f", summary.totalCost)}",
-                                style = MaterialTheme.typography.titleSmall,
-                                fontWeight = FontWeight.Bold,
-                                color = MaterialTheme.colorScheme.secondary
-                            )
-                        }
-                        
-                        // Display usage by model if there are models
-                        if (summary.usageByModel.isNotEmpty()) {
-                            Spacer(modifier = Modifier.height(24.dp))
-                            
-                            Text(
-                                text = stringResource(R.string.usage_by_model),
-                                style = MaterialTheme.typography.titleSmall,
-                                fontWeight = FontWeight.Medium
-                            )
-                            
-                            Spacer(modifier = Modifier.height(8.dp))
-                            
-                            // Top 3 most used models
-                            val topModels = summary.usageByModel.values
-                                .sortedByDescending { it.totalTokens }
-                                .take(3)
-                            
-                            topModels.forEach { modelUsage ->
-                                Row(
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .padding(vertical = 4.dp),
-                                    horizontalArrangement = Arrangement.SpaceBetween,
-                                    verticalAlignment = Alignment.CenterVertically
-                                ) {
-                                    Column(modifier = Modifier.weight(1f)) {
-                                        Text(
-                                            text = modelUsage.modelName,
-                                            style = MaterialTheme.typography.bodyMedium,
-                                            fontWeight = FontWeight.Medium,
-                                            maxLines = 1
-                                        )
-                                        Text(
-                                            text = "${modelUsage.totalTokens} tokens",
-                                            style = MaterialTheme.typography.bodySmall,
-                                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                                        )
-                                    }
-                                    
-                                    Text(
-                                        text = "$${String.format("%.4f", modelUsage.estimatedCost)}",
-                                        style = MaterialTheme.typography.bodyMedium,
-                                        fontWeight = FontWeight.Medium,
-                                        color = MaterialTheme.colorScheme.secondary
-                                    )
-                                }
-                            }
-                            
-                            // Show "View all models" button if there are more than 3 models
-                            if (summary.usageByModel.size > 3) {
-                                TextButton(
-                                    onClick = onViewAllModelsClick,
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .padding(top = 8.dp)
-                                ) {
-                                    Text(stringResource(R.string.view_all_models))
-                                }
-                            }
-                        }
-                    }
-                    is UsageViewModel.TokenUsageSummaryState.Empty -> {
-                        Column(
-                            horizontalAlignment = Alignment.CenterHorizontally,
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(vertical = 16.dp)
-                        ) {
-                            Icon(
-                                imageVector = Icons.Default.SearchOff,
-                                contentDescription = null,
-                                tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                                modifier = Modifier.size(40.dp)
-                            )
-                            Spacer(modifier = Modifier.height(8.dp))
-                            Text(
-                                text = stringResource(R.string.no_usage_data),
-                                style = MaterialTheme.typography.bodyMedium,
-                                textAlign = TextAlign.Center
-                            )
-                        }
-                    }
-                    is UsageViewModel.TokenUsageSummaryState.Error -> {
-                        Column(
-                            horizontalAlignment = Alignment.CenterHorizontally,
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(vertical = 16.dp)
-                        ) {
-                            Icon(
-                                imageVector = Icons.Default.Error,
-                                contentDescription = null,
-                                tint = MaterialTheme.colorScheme.error,
-                                modifier = Modifier.size(40.dp)
-                            )
-                            Spacer(modifier = Modifier.height(8.dp))
-                            Text(
-                                text = stringResource(R.string.usage_data_error),
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.error
-                            )
-                            Spacer(modifier = Modifier.height(4.dp))
-                            Text(
-                                text = tokenUsageSummaryState.message,
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Formats a Date as a relative time string (e.g., "5 minutes ago").
- */
-private fun formatLastUpdated(date: Date): String {
-    val now = System.currentTimeMillis()
-    val time = date.time
-    val diff = now - time
-    
-    return when {
-        diff < 60_000 -> "just now"
-        diff < 3_600_000 -> "${diff / 60_000} minutes ago"
-        diff < 86_400_000 -> "${diff / 3_600_000} hours ago"
-        diff < 604_800_000 -> "${diff / 86_400_000} days ago"
-        else -> "${diff / 604_800_000} weeks ago"
-    }
-}
